% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{comp_tau_mu_dev_pvals}
\alias{comp_tau_mu_dev_pvals}
\title{Compute Deviance and P-values}
\usage{
comp_tau_mu_dev_pvals(
  data.tbl,
  mu.vec.tst,
  tau.vec.tst,
  sm,
  y.k.in = NULL,
  sigma.2.k.in = NULL
)
}
\arguments{
\item{data.tbl}{A data frame containing the study data}

\item{mu.vec.tst}{Vector of mu values to test}

\item{tau.vec.tst}{Vector of tau values to test}

\item{sm}{Summary measure ("RR" or "OR")}

\item{y.k.in}{Optional pre-calculated effect sizes}

\item{sigma.2.k.in}{Optional pre-calculated variances}
}
\value{
A list containing the deviance matrix and p-value matrix
}
\description{
Calculates the deviance and p-values for the bivariate model
}
