% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{compute_confidence_region}
\alias{compute_confidence_region}
\title{Compute Confidence Region}
\usage{
compute_confidence_region(pval.mat, level.ma)
}
\arguments{
\item{pval.mat}{Matrix of p-values}

\item{level.ma}{Confidence level for meta-analysis}
}
\value{
A list containing the mu and tau values in the confidence region
}
\description{
Calculates the confidence region for mu and tau
}
