% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{inv_log_odds}
\alias{inv_log_odds}
\title{Inverse Log Odds Function}
\usage{
inv_log_odds(theta)
}
\arguments{
\item{theta}{A log odds value}
}
\value{
The probability corresponding to the input log odds
}
\description{
Calculates the probability from log odds
}
