% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{log_odds}
\alias{log_odds}
\title{Log Odds Function}
\usage{
log_odds(p)
}
\arguments{
\item{p}{A probability value}
}
\value{
The log odds of the input probability
}
\description{
Calculates the log odds from a probability
}
\details{
Helper Functions for Bivariate Meta-Analysis

This file contains helper functions for calculating log risk ratios,
log odds ratios, and their variances.
}
