% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{log_or}
\alias{log_or}
\title{Compute Log Odds Ratio Statistics}
\usage{
log_or(data.tbl)
}
\arguments{
\item{data.tbl}{A data frame containing the study data with columns:
event.e, n.e, event.c, n.c}
}
\value{
A list containing the log odds ratios and their variances
}
\description{
Calculates the log odds ratio and its variance
}
\examples{
data <- data.frame(event.e = c(10, 15), n.e = c(100, 120), 
                   event.c = c(5, 8), n.c = c(100, 110))
result <- log_or(data)
}
