% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{variance_log_rr}
\alias{variance_log_rr}
\title{Calculate Variance of Log Risk Ratio}
\usage{
variance_log_rr(event.e, n.e, event.c, n.c)
}
\arguments{
\item{event.e}{Event count in experimental group}

\item{n.e}{Sample size in experimental group}

\item{event.c}{Event count in control group}

\item{n.c}{Sample size in control group}
}
\value{
The variance of the log risk ratio
}
\description{
Calculates the variance of the log risk ratio
}
\examples{
variance_log_rr(10, 100, 5, 100)
}
